reorganize in-memory database code
